+++ libnd4j/include/loops/cpu/broadcasting.hpp
@@ -48,10 +48,9 @@ void Broadcast<X, Y, Z>::execInverse(int opNum, const void *x, const sd::LongTyp
 template <typename X, typename Y, typename Z>
 void Broadcast<X, Y, Z>::exec(int opNum, const void *x, const sd::LongType *xShapeInfo, const void *y,
                               const sd::LongType *yShapeInfo, void *z, const sd::LongType *zShapeInfo,
-                              long long int *dimension,
-                              int dimensionLength, const sd::LongType *xTadShapeInfo, const sd::LongType *xTadOffset,
+                              long long int *dimension, long long int dimensionLength, const sd::LongType *xTadShapeInfo, const sd::LongType *xTadOffset,
                               const sd::LongType *zTadShapeInfo, const sd::LongType *zTadOffset,
-                              sd::LoopKind::Kind loopKind, uint64_t start, uint64_t stop) {
+                              sd::LoopKind::Kind loopKind, long long int start, long long int stop) {
   DISPATCH_BY_OPNUM_TTT(exec,
                         PARAMS(x, xShapeInfo, y, yShapeInfo, z, zShapeInfo, dimension, dimensionLength, xTadShapeInfo,
                                xTadOffset, zTadShapeInfo, zTadOffset, loopKind, start, stop),
@@ -334,7 +333,8 @@ void Broadcast<X, Y, Z>::execInverse(const void *vx, const sd::LongType *xShapeI
                                      const sd::LongType *yShapeInfo, void *vz, const sd::LongType *zShapeInfo,
                                      sd::LongType *dimension, sd::LongType dimensionLength, const sd::LongType *yTadShapeInfo,
                                      const sd::LongType *yTadOffset, const sd::LongType *zTadShapeInfo,
-                                     const sd::LongType *zTadOffset, uint64_t start, uint64_t stop) {
+                                     const sd::LongType *zTadOffset,
+                                     long long int start, long long int stop) {
   auto x = reinterpret_cast<const X *>(vx);
   auto y = reinterpret_cast<const Y *>(vy);
   auto z = reinterpret_cast<Z *>(vz);
